Update 2020-12-17-an-introduction-to-session-types.md
On the web page where this is rendered, the latex expressions are put on a single long line.
That causes the comment signs '%' to comment out everything after it, and making
the latex not show as intended.  
The fix proposed here is to just remove the comment signs. 
An alternative is to not have the latex code turned into a single line,
but I don't know where in the process this happens.
Thanks for a great blog!